Test script. Updates RSS database.
I created this test script to invoke the newsdownld executable. The program is provided with parameters that directs the program to use XML files already downloaded. Each file is parsed to incorporate the contents into the database. The script was initially created as the primary means to test revisions to the XML parse routines that convert XML to rss data types. However, the script can serve as the basis for other scripts that use different parameters.