Show article summary data.
UI shows article summary information on the right side of the screen. The display widget is shown atop the article details box.

HTML detection is used to prevent html encoded data from appearing in the summary field. If the summary data did not populate in the database, then the article's headline is used in its place.

It is possible for the article details to be blank upon insertion in the database. This can happen based on the specific RSS format used. As a result, a filter is applied in the headline_view_select_row callback function in app_win.cpp. The modified code will now show information in the HTML details even when the database has no information in that field.