Manage RSS Feeds. A graphical UI window to facilitate adding/removing feeds.
Version 6 of the RSS program had the ability to add feeds. This new version improves on that by placing that functionality in its own graphical UI window.

The ability to add feeds was a feature of version 6 but that was achieved by placing that functionality on the same screen as the listed feeds. Although this was compact and convenient, it took away room on the main screen to show articles.

In that design, the feed name/url addition functions went unused most of the time. The other issue was that there was no mechanism to remove feeds.

This window in version 7 will consolidate feed management into a single screen separate from the main articles screen. That will lead to better expression of feed management functions.

**** This commit does not include any functionality. All it does is provide initial data entry fields and a few radio buttons. The window is still in a state of refinement.