Added button to refresh feeds.
A feature missing from version 6 was explicit feed refresh. The refresh feature existed but only implicitly when you clicked on the relevant feed button. This new button in version 7 will make the operation explicit so that the end-user can be sure that feed refresh was initiated.

Separately, updated the Manage Feeds button so that the callback function name matched in terms of plurality.