Consolidation + feed interval
Grouped all code statements for creating the window into a function.

Grouped all code statements for creating the feed input fields + layout into a function.

Added a field for the entry of the feed refresh interval. The refresh interval is the time when the feed update API is allowed to request new feed data from an Internet service. This is *extremely* important and is the one field that ensures the feed program works reliably.

If you request feed data from Internet hosts *too often* they will block you from downloading new feeds. The minimum refresh interval tends to be 1 hour if you are an unregistered user-agent (that applies in this case). A maximum of 4 hours as the minimum refresh interval is also configured into the interval input field. That could also be set to 24 but 4 seemed suffient as the longest time (in hours) to wait until a feed refresh solicitation is allowed.