Added main screen width/height.
The function definition for show_dialog now includes width/height passed in from the main screen. This will provide the RSS Management screen with the information needed to dynamically size itself in the context of the main screen. Since the RSS Management screen is configured as a non-resizable modal dialog the values passed in will remain the most accurate values until the RSS Management window is closed.