Retain the window's width/height at any given time either changes.
GTK/GDK has many functions in the API for getting a Window's width and height. I tried 4 or 5 different parts of the API to include functions for obtaining the size of the window. None of those API functions worked.

The reason I needed the window size is so I could determine a good default width/height for spawned dialog windows. The default of full screen was not my preference. I started simple with the GTK window_get_size function. The description for this function is simple enough. Unfortunately, it does not work except as a state value container to restore a previous window size.

When a program first starts up, the window's size is either -1 (for default sizing) or 0. Calling window_get_size then will result in this previous value of 0. The name of the API function does not reflect what people would intuitively gleen as to the behavior of the function. The documented description for the function does not make that clear. You have to get clarification when reading about window sizing in general.

Next I tried various API functions to include those that solicited the geometry of the window. Some of the geometry functions for a GTK window are deprecated. Later, I dug into GDK which underlies GTK to obtain a window_type_hint but what I really needed was a GTK Window Attribute structure that included width and height. Unfortunately, GTK Window Attributes obtained from GDK are blank as well.

I also tried the direct functions of window_get_width and window_get_height. Those did not work either. They also returned 0 on a valid window pointer to a window that has been made fully visible and running.

Finally, I tried the size-allocate signal off the Window. I thought about doing this first but wanted to try the other methods since they were real-time and thus more likely to be accurate at any given time. However, size-allocate was the right way to go after I determined the type definition for GdkRectangle by grepping GTK examples from git.

As a result, when the window is realized on screen, the actual height and width it has when a dialog is spawned from it is accurate at that time.