GTK Tree View to shows RSS feed configuration.
Updated the layout for the RSS Management screen. Added a GTK Tree View to show RSS information about the feeds, feed names, and access date. This will later be expanded to allow modification of the RSS feed configuration shown.

The RSS Management screen is initially sized in the context of the main RSS screen. This is accomplished by passing in the most recent window size prior to creating the RSS Management screen. See the commit details for app_win.cpp for details.