Need only the close button on dialog.
New constraints on the window removes the minimize button. Removing this button prevents a usability issue. Minimizing a modal dialog is less useful than keeping the dialog visible while active. In some situations, a minimized dialog may be an unexpected behavior. The actions needed to restore the dialog may be less obvious. The new constraints preserve the expected behaviors of a sub window that remains visible while in effect.

GDK Window Type Hint = DIALOG was used to affect this behavior. The default size of the window in the absence of valid values from the main window was adjusted to the expected minimum screen size.