More compact visual representation of retention option.
Merged individual radio buttons into a single combo box.

A single combo box, in this case, was a better choice visually. This resulted in less space needed on screen and made the presentation of retention options more compact.

A row for visually hoisting the radio buttons was removed that was no longer needed.