Expand RSS writer wit retrieve limit hrs and retention days.
The database will be redefined to hold additional values in the feeds table.

retrieve_limit_hrs

The retrieve_limit_hrs is to be used by the various RSS applications to prevent too frequent access to network servers. Network providers of RSS feeds usually have policies in place to prevent DDOS distributed denial of service. Acessing servers too often can result in a client endpoint getting blocked. This flag can be used by the application in conjunction with date/time API to determine when it is proper to access a network service.

retention_days

The retention_days help the application maintain the database but removing items declared stale. This flag would be used with date/time API to determine which items should be removed from the database. This can also be a trigger to use SQLite vacuum functions to reduce database size on disk.

*** Additional note ***
COMPATIBILITY

This change is not compatible with the existing applications. This interface change is a breaking change within the in-development version of the program. At the point of this commit, the code will not compile until the other applications are resolved to the API expansion.