Temporal method to decide when to download data.
Third-party networks do not want bots to access their sites too frequently. The application needs a way to know when it is too soon to access a third-party server hosting content. The is_stale function determines that in conjunction with meta-data in the database.

Each area of code that calls the download_rss_feed function is surrounded by the is_stale function. That way, the download_rss function is not called unless the time is right. The time guard is the sole means to achieve this.