Provide rowid for feed configuration based on url.
Principle function for the application to modify the exact row for a feed configuration. A feed configuration's url is unique within the entire system. However, during feed configuration changes, even the feed url can change. During a feed change operation does a requirement exist to use a mechanism other than the feed url to modify that configuration.

Other uses for the row id may arise in the future. The function was made since row id is not maintained in the feed data structure in the application.

This is not a software system that relies on any particular database. The implementation details of the rss/impl/reader.cpp and rss/impl/writer.cpp can be defined to use any storage broker. Currently, it uses a database but it could be redefined to use direct file I/O or network server.

Further, this system is oriented toward read-only data. The majority of information is downloaded, kept in storage and shown on screen. For example, the RSS feed headlines, the articles are never modified by the end-user. The only data for an article that could change is the publication date but this system keeps the original date and ignores publication date revisions in order to track when the article was first downloaded.

Since the body of information in the system is read-only, it is unnecessary to maintain conventional numeric primary-foreign keys as they will go unused. This also simplifies the implementation. The get_row_id function then provides the primary key for a table only when necessary. Last, since RSS configurations do not change often, this is further reason to keep database implementation details out of the data structures and APIs.