Maintain row selection following UI table reset.
When data is posted to the database, a query brings back the new state of the database. The UI table is refreshed. Any row that was visually selected is no longer selected.

Bringing back all the data makes the UI table match the state of the database. Multiple changes to the database are possible. Fresh data retrieval of all rows ensures what is shown on screen matches the database. That includes successful and failed attempts to change the database.

Reapplying data to the UI table undoes any visual selection indicators. This change re-establishes the visual selection indicator following a UI table update. Since the data entry fields are also updated when the row a row is selected, this ensures those fields match the current state of the database at the time the update button was pressed.