Pause before download.
Flag added that toggles activation of code that will suspend execution of the download_feed function. This provides time during testing to inject SQL to simulate new downloads. That is needed as a convenience during testing to properly test the process_feeds function. It is a convience as it avoids the need to setup a web server or invoke a web server like facility during testing. A web server is still recommended for more intense testing.