Render headlines by remapping widget to the screen.
A problem occured adding headlines to an existing list box widget. The program would abruptly terminate. This did not happen consistently but the source of error was the same in all cases. More lines of data added to the screen than can be restyled individually.

The new approach is to momentarily hide the list box and then show it again. The list box is updated while hidden. Once the update is finished, the list box is shown but with additional lines of data. This is equivalent to a batch update on the UI.

See commit 124263c5df3e3a35650194eaacf250c59b0b6a76 for details. The issue was originally thought resolved in that commit. More extensive testing revealed the issue occured when network timing varied along with a larger data set. This commit is associated with an expanded testing regime. The UI is solid at this point on the matter of concurrent updates to the screen from a background thread involving numerous updates to multiple tabs.