New feeds appear as new tabs in real-time.
When a new feed was added, the data would reside in the system correctly but the screen did not properly reflect the s
tate of the database. Adding a feed would not result in a new tab.

The program has been tested several times to assert that a tab is created whenever a new feed is added. Worked on this one for about 6 hours straight. Root cause described in commit 5888945127faf533db95f5bda89a1216388ee784.

The set of changes introduced to better express the solution has more of a production feel to it. The code was reduced while working towards a better result.