Extracted tab initialization into a function.
Tab initialization occurs when the window is being built. The initialization code for the tabs is more readable in its own function.

During the process of setting up the separate initialization function, I attempted to invoke it from a separate thread. That conflicted with the asynchronous nature of UI kit composition. Particularly in the CSS engine. The initialization function was then readapted to be called from the site where the headlines_view variable is defined.

I identified an interval of about 2 - 4 seconds for complete CSS engine initialization in relation to threaded composition of a UI using this kit. Given the duration, I decided a synchronous function call from within the create function would be just as quick, more reliable, and easier to understand. The optimization opportunity weighed less than the maintenance effort to sustain programmer directed asynchrony at the UI level on start-up.

Next, I identified conflicts in object deallocation in the UI kit related to timing of thread conclusion. UI elements originating from a non-primary thread may either double-free or otherwise fail internal consistency checks resulting in SIGSEGV or SIGABT. The matter concerned automatic hierarchical deallocation in a container tree. When threads are involved, this can operate in a non-deterministic way resulting in signal abort related to the CSS engine or indeterminate relations among contained elements and their containers.  The use of g_object_ref_sink function mitigated the latter issues fully. I left in place the use of g_object_ref_sink and the corresponding gtk_widget_destroy calls as that seems to have resolved other issues related to concurrent UI operations.

The end result is the tab initialization is not functionally different. Rather, the code is better aligned to the characteristics of the UI kit.