Provides actual class name of a GTK Widget.
Extremely useful function. Save hours debugging, many minutes in documentation.

GTKWidget is generic. Not just generic, but ultra generic.

The highly generic nature of GTKWidget maximizes reusability.

A common, abstract interface like this exposes numerous widgets to more functions. The downside is that sometimes when you swap out widgets, the replacement widget does work the same way. Updated code will sequence properly but a gap in interface requirements means the widget may not show any information.

An example is the difference between GTKListBox and GTKTreeView. Both of these visual components functions well with scrollbars. The scroll bars are not automatic and you need a GTKScrolledWindow widget to manage the scroll bars. The problem is when you switch from GTKListBox to GTKTreeView, the way you get back a pointer handle to either a list box or tree view changes.

The conditions are noted in the documentation for GTKViewort.

GTKViewport is listed under the "Miscellaneous" section of the GTK 3.2x series documentation. Under the description section of the GTKViewport page, it states that viewport will be added "in between" the actual widget and the scroll window. However, some widgets do not need this including GtkTextView, GtkTreeView or GtkIconView.

What that means is if you have psuedo code such as:

	obj = gtk_bin_get_child (scroller)

You will either get the viewport or the underlying widget depending on what widget was loaded into the scroll window. Not a problem though because the function in this commit tells you the actual type name of "obj" in the example above. You just fork the code path accordingly.