Exclude RPM and SRPM directories.
A build script pulls the RPM and SRPM directories to the local machine. The directories originate on a build server configured to produce source and binary RPMs. These archives are not intended to be revision controlled but pushed from the local machine to test VMs to validate the configuration of RPM packages defined by the newsreader.spec file.