Red Hat suggested security reinforcement in the code.
Red Hat describes compiler flags that enables precautions involving code security. This commit adds those flags.

The program was tested with those flags applied. No observable difference in performance. Correctness and maintainability is the goal with this project rather than performance. Yet, it is still good to see no regression in code execution time.

I will have to figure out what to do with the linker option, -flinker-output=pie. That is not expected to work with Ubuntu in terms of double-click launch of the executable in that operating system. A workaround for Ubuntu make exist in the form of a launcher. That is an issue I will work through later. On Fedora/Red Hat/CentOS, the option is suggested as part of an overall secure profile for the program when it comes to code to exe production.

See the document from Red Hat:
	Developing Applictions in RHEL 8 (2019-11-15)