Modify UI from non-UI thread.
Second step in resolving improper access to the UI from a non-UI thread. Note that removal of runtime assertions is not part of the solution. The effect of removing the assertions only mitigates against abrupt shutdown of the program. However, the conditions that result in the assertions raising the abort signal persist without the approach taken in this commit.

A few, "very acute", critical sections are defined in the code to modify the UI from a non-UI thread. The approach is more like using a scapel to implant the appropriate thread entrance/exit routines to minimize blocking on the UI. I tried a variation that involved setting a persistent thread enter call at the start of the application or within the second thread loop. Both approaches blocked the UI too much and so the scapel approach was taken. That results in more code, but the application runs better.

See the following commits for details:

781ea75dfe4a6ec62d77fefcba2093364fdb9988
092b8e40f61079e2888998a5076cd5a7fbbe22ee