License clarification, argtable.
Fedora Packaging Guidelines state that RPM packages identify governing licenses. Meeting the guidelines is more straightforward if a single license applies to all source code. Provisions exist to declare multiple licenses. I am building an RPM definition that can be applied to CentOS, Fedora, and Red Hat Enterprise Linux. That RPM definition applies only to the newsreader program which is wholly licensed under LGPLv2+.

The source repository the project draws from includes a copy of argtable.h and argtable.c which is governed under a license compatible to GPL and LGPLv2+. However, since the argtable project or the version of argtable included in the repository does not explicitly declare an OSI license, I am excluding it from the RPM build source tree (as defined in the script rpmbuild.sh under package_build/) to avoid any confusion.