Source RPM for newsreader derived only of source files needed to build executable version.
All files in source RPM governed under LGPLv2+ and the resultant binary will be LGPLv2+ in observance of suggestions found in the Fedora Packaging Guidelines (Licensing Guidelines section). The previous iteration was acceptable. This version is more exact in the inclusion of files. The license tag in the newsreader.spec file is now fully applicable to all files in the source RPM and the files in the derived binary RPM.