Prepare for tailored install steps.
A discrepancy exist in the RPM packager documentation including the various guides on the Internet.

Most guides suggest using the RPM macro %{make_install}. Unfortunately, the default values for %{make_install}, at least on Fedora 31 server (operating as a build server) does not supply the correct value for DEST dir that immediately translates into the appropriate runtime destination directory. The value provided to /usr/bin/install is /usr/local/bin which violates the checks from rpmlint.

As a result, this spec file defines translations from the staged install directories under RPM %{buildroot}/usr/local/bin to %{buildroot}/usr/local which corresponds to the destination expected by rpmlint. That satisfies the checks for rpmlint for directory structure. The file app.css will need to be placed in a data directory however.

I chose this solution instead of manually defining the invocation of both the make and install programs. The reason is I want to maintain the suggested parameters from the rpm environment supplied to make and install. I copy the files from the staged installation directories as a means to reflect the state of the files in those directories in the required runtime directories.

I then remove the staged install directories (as in: %{buildroot}/usr/local/bin ) since rpmlint does not like packaged files in 2 places. I significant change is coming to the spec file in the install and files sections. I will add a documentation file and app.css needs to go to a new location per rpmlint.