Future newsreader spec file planning.
The spec file for the newsreader program needs more information. As configured, the newsreader spec file will result in an RPM that either does not install at all or install improperly. The missing pieces are as follows:
- No Linux manual page.
- No screen shots.
- No desktop file.
- No operating system menu integration.
- Improper placement of data and non-executable dependencies.

The last 2 items on that list are the most important. While I am doing that however, I think it is a good idea to address the other items as well.

I saw a good example of a finished %files section at https://linuxconfig.org/how-to-create-an-rpm-package. I found the article in an attempt to resolve a different issue. While the article did not apply to my scenario, I realized the utility in the example under the "Create the specfile" section of the article.

This spec file in this commit will have placeholders on lines 49 - 61. I will erase the comment markers once the actual materials are available and correctly cited in a future RPM build.