Annotate RPM spec file.
Note about %{make_install} when applied to rpmbuild program on Fedora 31 server. When using Fedora 31 server as a build server with rpmbuild installed, the execution of the RPM macro %{make_install} does not correlate to the value expected by rpmlint for %{_bindir}. This appears to be a bug that will require a permanent workaround as denoted in the spec file in this commit. See lines 61 - 64 as the workaround applied.