Annotate RPM spec file. Finalized for exe install.
This version of the spec file is good to go for the install of the newsreader executable but only that file. The app.css file still has to be reworked in terms of a suitable installation directory. Also working on documentation. I am going to have to learn man pages language from scratch in the next hour or so ...

In reference to commit f555f7956c5c4d32197fae9993936e2040cf4a8d, the workaround is actually in the spec file (see the present commit, lines 43 - 52) are acceptable. As mentioned in commit f555f7956c5c4d32197fae9993936e2040cf4a8d, I perceived the discrepancy between the default use of %{make_install} and the resultant interpretation in rpmlint as a bug. An article on UNIX StackExchange shows that this is actual an intentional design choice under Fedora.

The link, https://unix.stackexchange.com/questions/125120/why-is-dir-or-file-in-usr-local-an-error-rather-than-a-warning?rq=1 implies that Fedora system policy defaults to validating use of /usr/local/bin as an invalid value through rpmlint. However, the default in rpmbuild is to produce a build output in /usr/local/bin. I decided to accept the default situations in both rpmbuild and rpmlint and simply customize the process in-between (my spec file applied to rpmbuild).

I think this is the intended approach. The growing consensus is that rpmlint is a good tool but is not authoritative when it comes to the status of an RPM you are setting up. I decided to continue using rpmlint as I think it is valuable, but in this one area, I am applying an override in the rpm spec file to designate the program executable to /usr/bin instead of /usr/local bin. On my own Fedora system, no single program I've installed from the official Fedora repository or RPM fusion has gone into /usr/local/bin. I understand the concept of /usr/local/bin, but it the overall sense is it is best used for in-house programs on local area networks rather than those sourced from public networks.