Updated build to v7.0.8.
Tested the build process with a live connection to github. The rpmlint program validates the package with no errors or warnings.

The package is not yet tested on an actual desktop system. A virtual machine has been allocated for testing.

The RPM package that is built is currently 100% in terms of specification and pre-install validation.