Add appdata.metainfo to RPM.
Fedora Packaging Guidelines suggest an appdata file (see also AppStream).

The newsreader.appdata.xml describes the program. This description would be shown in a software center (aka Linux style app store).

The appdata file is packaged according to the Fedora Packaging Guidelines. No screenshots exist. The file will be updated once screenshots are available.

This is the start of an appdata.xml file. A version that has screenshots will coincide with a later release. Probably release 7.10.