Situate rss data in a per-user directory.
The Linux File Hierarchy Standards contains a provision for per-user data. As part of an effort to orient the program to FHS, I reworked program initialization to place data unique to the end-user in a standardized directory tree under the user's home directory in Linux. That way, when the program is installed to the global /usr/share/* locations, actual RSS data remains specific to a user account.