Fedora / Red Hat packaging .spec file simplified.
See commit: 1d124accc2e6eff1af31eee640bc8a5a605d5d3d

The previous version had numerous workarounds for a design flaw introduced in the default settings of bakefile. After seeing the same issue appear when building a .deb package for Ubuntu, I realized, during the course of changing $PREFIX in the bakefile, that the bakefile default was incompatible with package management in Red Hat/Debian worlds.

This .spec file is substantially reduced in size. The number of instructions needed to build the software install package is minimal. More of the automatic methods in rpmbuild can now be used productively. It is good to know however that they can be overridden.

Previous commits notes where I speculate about errors in the rpmbuild tools were mistaken. The root cause was identified and addressed.