Performance optimization planned. I tried the "out-of-the-box, maximum readability" approach. That does not work 100% of the time, but I had to look into it. The program has an excellent foundation but needs a performance tune-up. There are just some things a compiler does not do better than a human.
No code changes are in this commit. The information here are notes on what I am about to change.

The program accesses a database on my machine with the following record totals.

 Ars Technica           443
 BBC                  1,327
 DistroWatch             15
 Linux Kernel            32
 MSN                  3,039
 Phoronix               411
 Slashdot               489
 Tech Rights            214
 TechCrunch             613
 The Atlantic           540

The size of the database file is now 24MB.

The average length of headline data is roughly 40 – 70 characters per record. The data is not large by enterprise data standards. However, the result of loading this small database under the current processing sequence is about 25 – 30 seconds. The slow launch time is noticeable and unacceptable.

I conducted tests using the following command sequence:

 date && ./newsreader && date

I pressed ALT+F4 once the screen appears.

The command sequence captured the full cycle of activity from an end-user standpoint. The time overall is constant at 25 seconds. At least that is what the output on the command line read. The "perceived time" seems longer. You double-click on the program's icon and you are just waiting for the program to open.

I then adjusted the database to remove the feed with the most records (temporary removal). That gained me 10 seconds launch time.

Each feed shown in the list above is represented as a tab in the user interface. During program initialization, the program automatically loads each tab with a list of headlines for that feed. This occurs before the initial screen is rendered. The list of headlines is held by the respective tab object's listing object. When the user interface window is shown, the tabs are also shown. Code is executed on every headline entry increasing the time it takes for the tabs to report that they are ready to be shown on screen. Part of the solution is to take away unnecessary start-up initialization. The core of upcoming revisions will be to have tabs load their content on demand and in a way that is visually responsive.

Preliminary analysis is that performance reduction is influenced by but does not occur on line 337, 679 and 691. These specific lines of code in app_win.cpp only sets the conditions for greatly lowered performance. The performance is actually reduced when the call to render the window occurs on line 362.

The goal is change the performance of the program so it presents any locally available data in a more responsive fashion. The solution will involve gutting out lines 337, 679 and 691 in their present forms. This will impact not only lines 335 – 337, but also lines 670 – 713.

The final solution will work in a substantially different way. This includes how or if data is loaded on start-up. This will include significant modifications to the way data is loaded when a tab is clicked.

The amount of RAM used in the update will actually decrease significantly but only as a side-effect of how the solution will be implemented. This computer program will see a huge increase in "code complexity" to provide a smoother user interface interaction. A data load buffer will be introduced that will control what data is actually presented rather than leave it solely to the GTKNotebook object. The program will conduct fine grained control over the “switch-page” signal handling provided by the GTKNotebook object to determine which tabs have data and in what way.