Achieved sub-second load time. - Perceived load time of the program will appear to be instantaneous. - Faster rendering of tabs. The first tab shown is fully rendered on load.
Approach:

The best technique with this type of user interface is to truncate the initial data per tab. As an example, a tab may represent 10,000 items saved to the database. Not all 10,000 items need to be displayed at once. In fact, the nature of this program informs a representation that favors progressive loading of data. The data fed into tabs will be reworked in subsequent updates. The data flow will include a temporal sequencing element in which items will appear from most recent to least. That will make the progressive loading of data in a given tab more useful though the principle side benefit will be higher performance.

Once I have the right implementation in place, the program will be able to handle thousands of feed entries in the first 1 - 2 minutes after the program starts. This release also includes a switch from the gthreads API to the GDK thread idle functions. Those functions are not as good as granular thread management but the new code is closer to best practice in GTK. The thread idle function callbacks work but it remains to be seen how well they will work in terms of multiple GDK thread idle callbacks.

Another part of the performance plan is to decouple feed download processing from user interface update. The code is currently defined to download and update the user interface at the same time. Although this is convenient, it will conflict with the new data flow I have put in place. A decoupled download process will work better architecturally long-term. More code will be needed to fine tune the detection of new data at the correct point in the pipeline. I will reconcile the retrieval of data in the process of loading into tabs and data coming from the network that needs to be loaded as well.
- Plan is to progressively load data in each tab.