RSS feed downloads within a separate thread.
The download process operates independent of the user interface. No part of the user interface is updated by the download process. This provided an opportunity to make the download process more robust. The download process occurs in an explicit thread initiated from the UI. This provides more detailed control over the concurrent operation.

At the same time, I finally got around to implementing the download process the way I envisioned where excessive calls to the network are controlled in a *fine grained manner*. That is to say, the process no longer checks every few seconds if there are new feeds to download. Once downloads occur, the process then holds off on any further checks until 12 minutes later.

That has an immense impact on the output on the command-line if the UI is launched from a Terminal program. The output to stdout is more informative but not overdone. The cycle from download, wait, check, and download again when appropriate has all been tested. The implementation is in perfect order.

The other change is the use of int64_t. I was reading up on SQLite rowid for a technique I am about to pilot and noticed verbiage about the key size. I foresaw an issue and took steps to address it by using a data type in the application that corresponds to the one used by SQLite. I could have used [long long] but decided that int64_t seemed cleaner to express and potentially more efficient and predictable in use across architectures the program may eventually run on.