Provide a data range (new downloaded entries) to update UI.
The program downloads entries to the database in a separate thread. A best practice is to not update the UI from an explicit thread. This thread sets two fields with the begin/end range of records added since the program started. That signals to UI threads managed by GTK that new records are available. This approach fully decouples data download integration from UI processing. The UI will later retrieve the new records and add them to the UI.