Restored rss management implementation.
While the main separate thread implementation for rss feed download was in progress, the management functionality was turned off. Now that the primary functionality is in place, the rss management function has been re-implemented. The rss configuration changes still occur in a separate thread as well, but in a GTK managed thread. A single attempt to download feed entries occur when a new feed is added. Also, an attempt is made to synchronize configuration changes to the user interface.