Starting snapshot oriented configuration change.
The configuration screen managed feed meta-data by directly editing the database on each change. That worked well in the beginning but as the level of asynchronous operations increased in the main screen where feeds are viewed, that became less tenable. One solution is to modify the feed configuration in-memory first. The final modifications would then be released to the database and the main screen.