Hold main screen operations until configuration window exits.
When the main screen launches the configuration screen, some operations should be paused or discontinued. Once the main screen receives an indication that the configuration screen is finished, it can resume or start processes for switching tabs and downloading information. The configuration screen has a process for passing data that wsa initialized in the main screen. This commit includes the proper calls to the appropriate parts of teh configuration screen's interface.