Configuration snapshot load and modify.
The snapshot is successfully loads when the screen first appears. Minimal testing has been done with the updated modification process. Initial tests show that the modification succeeds in most areas. An issue exist when both name and url are modified at the same time. That will be addressed in a later commit.

An issue developed with the memory model. That was reflected in the handlers for the treeview after the signals executed. Abnormal ends coincided with access to treemodel iterators and entry buffers.
An indirect cause was the excessive use of stack pointers in C++ that impacted use of malloc in C. That led to an observation that it is far better to copy by default than have a call chain with values aliased to an original source. Once that was cleaned up, the memory model issues seemed to have disappeared.

Additional updates were formed from insight through examining how GTK itself uses gchar* and GValue. Access to values of that type from GTK API were appropriately modified in expression to enact proper initialization and deallocation. Although that did not resolve issues at first, after other changes, this appeared to stabilize program operation in the configuration screen.

Other substantial changes are planned following this commit.